Add OBS Websocket advanced module
I have just rewritten from scratch the original OBS module, the purpose was to be more programatical about the protocol: to avoid manual labour, and manual methods on everything, stuff are generalized to parse the original api definition and transform it into Chataigne's world.

My hope is that it is more robust this way, and easier to maintain as well. If the API changes it just needs to be regenerated.